+++ src/store/account/__tests__/mutations.spec.js
@@ -1,12 +1,12 @@
 import mutations from '../mutations';
-import { initialState } from '../index';
+import store from '../index';
 
 describe('store/account/mutations', () => {
   let state = {};
 
   beforeEach(() => {
     state = {
-      ...initialState,
+      ...store.initialState,
     };
   });
 
@@ -79,6 +79,16 @@ describe('store/account/mutations', () => {
     mutations.setMembership(state, membership);
 
     expect(state.membership).toStrictEqual(membership);
+  });
+
+  test('mutations.setBuyMembershipTx', () => {
+    const tx = {
+      id: 1,
+    };
+
+    mutations.setBuyMembershipTx(state, tx);
+
+    expect(state.buyMembershipTx).toStrictEqual(tx);
   });
 
   test('mutations.setRewards', () => {