+++ src/store/account/actions.js
@@ -13,6 +13,7 @@ export default {
     });
     await Promise.all([
       dispatch('fetchMembership', address),
+      dispatch('fetchBuyMembershipTx', address),
       dispatch('fetchBalances', address),
       dispatch('fetchDelegations', address),
       dispatch('fetchUnbondings', address),
@@ -47,21 +48,51 @@ export default {
     }
   },
   /**
+   * @param {Function} dispatch
    * @param {Function} commit
    * @param {String} address
    */
+  //TODO: add  dispatch param
   async fetchMembership({ commit }, address) {
     try {
       const response = await api.requestMembership(address);
       commit('setMembership', response.data.result);
     } catch (error) {
       commit('setMembership', null);
-      //TODO: enable to manage the error when the API will be fixed
+      //TODO: Enable when resolved CORS type error in devnet
       // if (error.response && error.response.status === 404) {
       //   commit('setMembership', null);
       // } else {
       //   dispatch('handleError', error);
       // }
+    }
+  },
+  /**
+   * @param {Function} dispatch
+   * @param {Function} commit
+   * @param {String} address
+   */
+  async fetchBuyMembershipTx({ dispatch, commit }, address) {
+    const query = `assign_membership.owner=${address}`;
+    try {
+      let response = await api.requestSearchTransactions({
+        query: query,
+        page: 1,
+        limit: 1,
+      });
+      const pageTotal = response.data.page_total;
+      if (parseInt(pageTotal) < 1) {
+        commit('setBuyMembershipTx', null);
+        return;
+      }
+      response = await api.requestSearchTransactions({
+        query: query,
+        page: pageTotal,
+        limit: 1,
+      });
+      commit('setBuyMembershipTx', response.data.txs[0]);
+    } catch (error) {
+      dispatch('handleError', error);
     }
   },
   /**