+++ src/store/account/getters.js
@@ -1,5 +1,6 @@
 export default {
   balances: (state) => state.balances,
+  buyMembershipTx: (state) => state.buyMembershipTx,
   delegations: (state) => state.delegations,
   error: (state) => state.error,
   isLoading: (state) => state.isLoading,