+++ src/store/overview/api.js
@@ -5,40 +5,42 @@ export default {
   /**
    * @returns {Promise}
    */
+  requestAbrTokens() {
+    return axios({
+      method: 'get',
+      baseURL: `${API.KYC}`,
+      url: `/funds`,
+    });
+  },
+  /**
+   * @returns {Promise}
+   */
+  requestVbrTokens() {
+    return axios({
+      method: 'get',
+      baseURL: `${API.VBR}`,
+      url: `/funds`,
+    });
+  },
+  /**
+   * @param {String} address
+   * @returns {Promise}
+   */
+  requestBalances(address) {
+    return axios({
+      method: 'get',
+      baseURL: `${API.BANK}`,
+      url: `/${address}`,
+    });
+  },
+  /**
+   * @returns {Promise}
+   */
   requestAllTokens() {
     return axios({
       method: 'get',
       baseURL: `${API.SUPPLY}`,
       url: '/total',
-    });
-  },
-  /**
-   * @returns {Promise}
-   */
-  requestNodeInfo() {
-    return axios({
-      method: 'get',
-      baseURL: `${API.NODE_INFO}`,
-    });
-  },
-  /**
-   * @returns {Promise}
-   */
-  requestPool() {
-    return axios({
-      method: 'get',
-      baseURL: `${API.STAKING}`,
-      url: '/pool',
-    });
-  },
-  /**
-   * @returns {Promise}
-   */
-  requestStakingParameters() {
-    return axios({
-      method: 'get',
-      baseURL: `${API.STAKING}`,
-      url: '/parameters',
     });
   },
 };