+++ src/views/overview/__tests__/OverviewBody.spec.js
@@ -1,0 +1,55 @@
+import { createLocalVue, shallowMount } from '@vue/test-utils';
+import Vue from 'vue';
+import Vuetify from 'vuetify';
+import OverviewBody from '../OverviewBody.vue';
+
+Vue.use(Vuetify);
+const localVue = createLocalVue();
+localVue.use(Vuetify);
+
+describe('overview/OverviewBody.vue', () => {
+  test('if loading message is displayed', () => {
+    const wrapper = shallowMount(OverviewBody, {
+      localVue,
+      computed: {
+        error: () => null,
+        isLoading: () => true,
+      },
+    });
+
+    expect(wrapper.find('[data-test="loading"]').exists()).toBe(true);
+    expect(wrapper.find('[data-test="error"]').exists()).toBe(false);
+    expect(wrapper.find('[data-test="content"]').exists()).toBe(false);
+  });
+
+  test('if error message is displayed', () => {
+    const wrapper = shallowMount(OverviewBody, {
+      localVue,
+      computed: {
+        error: () => ({
+          message: 'Error',
+          status: 400,
+        }),
+        isLoading: () => false,
+      },
+    });
+
+    expect(wrapper.find('[data-test="loading"]').exists()).toBe(false);
+    expect(wrapper.find('[data-test="error"]').exists()).toBe(true);
+    expect(wrapper.find('[data-test="content"]').exists()).toBe(false);
+  });
+
+  test('if content is displayed', () => {
+    const wrapper = shallowMount(OverviewBody, {
+      localVue,
+      computed: {
+        error: () => null,
+        isLoading: () => false,
+      },
+    });
+
+    expect(wrapper.find('[data-test="loading"]').exists()).toBe(false);
+    expect(wrapper.find('[data-test="error"]').exists()).toBe(false);
+    expect(wrapper.find('[data-test="content"]').exists()).toBe(true);
+  });
+});