+++ changelog.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## Version 2.2.9, 19/05/2021
+
+- News:
+  - Added new page "Token Exchange Rate"
+  - Added link from account details to buy membership tx
+
 ## Version 2.2.8, 30/04/2021
 
 - News: